Интерфейс Input [#181092 #395486]

+++ src/main/java/ru/job4j/poly/Transport.java
@@ -5,5 +5,5 @@ public interface Transport {
 
     void passengers(int passengers);
 
-     int  refuel(int fuel);
+     double  refuel(double fuel);
 }
